fix(globusauthenticationflow): add retry for confirmation
Occasionally, the Globus authentication flow breaks at the point of
confirmation. This is after being authenticated by Globus, after being
redirected to KBase's callback, but before being dropped into the target
narrative. KBase presents a confirmation button which we need to click.

Rather than spending more time trying to reproduce the conditions that
lead to this error (it's sporadic and there are several layers to the
DOM we're interacting with), this simple retry logic should (hopefully)
resolve the problem, assuming it's a timing problem.